Move PostalCodeProperties from core to card module; Replace unnecessary RestrictTo annotations with internal visibility modifier for enums in CardValidationUtils.kt
